backend/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg[binary]==3.1.12
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.1.0 